contest_5 4

+++ contest_5/src/IShape.cpp
@@ -1,15 +1,14 @@
-#include "IShape.h"
+#include "../IShape.h"
 
-namespace geometry
-{
-  
+namespace geometry {
+
 std::string IShape::StringFormat(std::string class_name, std::string args[], size_t length) const {
   std::string ret = class_name + "(" + args[0];
-  for (int i = 1; i < length; i++) {
+  for (size_t i = 1; i < length; i++) {
     ret += ", " + args[i];
   }
   ret += ")";
   return ret;
 }
 
-} // namespace geometry
+}  // namespace geometry
